Setting up sandbox events so that the sandbox always listens for
an event under a namespace. That namespace will be
moduleName-moduleName for its' own events

and
publishingSandbox-moduleName for events from other sandboxes.

That way, we can publish events between sandboxes while still
maintaining isolation.

The downside is that it double the number of events added when listening
for something, so this broke a couple of tests looking at number of
listeners. I'm trying to figure out the consequences of this right now.